💄 refactor: design signin page with shadcn-vue
