+++ components.d.ts
@@ -16,13 +16,11 @@ declare module 'vue' {
     CardTitle: typeof import('./inertia/components/ui/card/CardTitle.vue')['default']
     Checkbox_group: typeof import('./inertia/components/form/checkbox_group.vue')['default']
     Checkbox_month: typeof import('./inertia/components/form/checkbox_month.vue')['default']
-    Error: typeof import('./inertia/components/form/error.vue')['default']
-    Footer: typeof import('./inertia/components/footer.vue')['default']
-    Header: typeof import('./inertia/components/header.vue')['default']
+    Error: typeof import('./inertia/components/Error.vue')['default']
+    Footer: typeof import('./inertia/components/Footer.vue')['default']
+    FormInput: typeof import('./inertia/components/FormInput.vue')['default']
     Input: typeof import('./inertia/components/ui/input/Input.vue')['default']
-    Input_group: typeof import('./inertia/components/form/input_group.vue')['default']
     Label: typeof import('./inertia/components/ui/label/Label.vue')['default']
-    Message: typeof import('./inertia/components/message.vue')['default']
     NavigationMenu: typeof import('./inertia/components/ui/navigation-menu/NavigationMenu.vue')['default']
     NavigationMenuContent: typeof import('./inertia/components/ui/navigation-menu/NavigationMenuContent.vue')['default']
     NavigationMenuIndicator: typeof import('./inertia/components/ui/navigation-menu/NavigationMenuIndicator.vue')['default']
