+++ inertia/layouts/AppLayout.vue
@@ -1,0 +1,69 @@
+<template>
+  <div class="flex flex-col min-h-full">
+    <header class="sticky flex justify-between border-b border-border">
+      <NavigationMenu class="ml-7">
+        <NavigationMenuList>
+          <NavigationMenuItem>
+            <Link
+              href="/"
+              class="hover:text-muted-foreground hover:bg-background focus:bg-background"
+              :class="navigationMenuTriggerStyle()"
+            >
+              Plant Garden
+            </Link>
+          </NavigationMenuItem>
+        </NavigationMenuList>
+      </NavigationMenu>
+      <NavigationMenu class="mr-7">
+        <NavigationMenuList>
+          <NavigationMenuItem v-for="link in links" :key="link.name">
+            <Link
+              v-if="page.props.isAuth === link.auth"
+              :href="link.url"
+              class="hover:text-muted-foreground hover:bg-background focus:bg-background"
+              :class="navigationMenuTriggerStyle()"
+            >
+              {{ link.name }}
+            </Link>
+          </NavigationMenuItem>
+        </NavigationMenuList>
+      </NavigationMenu>
+    </header>
+    <slot />
+    <Footer />
+    <Toaster />
+  </div>
+</template>
+
+<script setup lang="ts">
+  import { watch, ref } from 'vue'
+  import { usePage, Link } from '@inertiajs/vue3'
+  import { navigationMenuTriggerStyle } from '@/components/ui/navigation-menu'
+  import { useToast } from '@/components/ui/toast/use-toast'
+  import { SharedProps } from '@adonisjs/inertia/types'
+
+  const page = usePage<SharedProps>()
+
+  const links = ref([
+    { name: 'Mes jardins', url: '/gardens', auth: true },
+    { name: 'Calendrier', url: '/calendars', auth: true },
+    { name: 'Plantes', url: '/plants', auth: true },
+    { name: 'Déconnexion', url: '/logout', auth: true },
+    { name: 'Connexion', url: '/login', auth: false },
+    { name: 'Inscription', url: '/signin', auth: false },
+  ])
+
+  const { toast } = useToast()
+
+  watch(
+    () => page.props.message,
+    (newValue) => {
+      if (newValue) {
+        toast({
+          description: newValue.description,
+          variant: newValue.type === 'error' ? 'destructive' : 'default',
+        })
+      }
+    }
+  )
+</script>
