+++ inertia/pages/main.vue
@@ -7,7 +7,7 @@
 </template>
 
 <script setup lang="ts">
-  import Layout from '@/layouts/default.vue'
+  import Layout from '@/layouts/AppLayout.vue'
 
   const props = defineProps<{ title: string }>()
 </script>
